net: introduce HW Rate limiting driver API
The network devices gain a new ops struct to directly manipulate the
H/W shapers implemented by the NIC.

The shapers can be attached to a pre-defined set of 'domains' - port,
device, etc. - and the overall shapers configuration pushed to the H/W is
maintained by the kernel.

Each shaper is identified by an unique integer id based on the domain
and additional domain-specific information - e.g. for the queue domain,
the queue id.

The shaper manipulation is exposed to user-space implementing
the NL operations described by the previous patch.

Note that an additional domain, not exposed to user-space, is defined
here (VF - to allow implementing the existing ndo_set_vf_rate() on top
of the API defined here).

Co-developed-by: Simon Horman <horms@kernel.org>
Signed-off-by: Simon Horman <horms@kernel.org>
Signed-off-by: Paolo Abeni <pabeni@redhat.com>
---
v4 -> v5:
 - fixed a few typos
 - set() operation commit changes in memory on success
 - updated set() to handle positive return values for device op
 - get() always fill the parent handle

v3 -> v4:
 - fix build error with !NET_SHAPER
 - fix skipping one item on shaper dump requiring multiple skbs
 - add a bunch of comments
 - clarified the return value for delete()
 - rebased on yaml changes

v2 -> v3:
 - completed the implementation for set/get/dump/delete NL ops
 - plugged shaper cleanup at device dismatel